EXData

Description

In module EXData we introduce two new data types(EXError, Expression) and describe what some existing types represents in our functions, so we are making synonims for already existing types.

Documentation

type Coordinates = (Int, Int)

The type of Coordinates(Coordinates represent cell coordinates) is (Int, Int).

type References = [Coordinates]

The type of References, which are references for other cells in the expression, is [(Int, Int)].

type FeedbackValue = (Coordinates, Maybe Integer)

Type FeedbackValue holds information about one cell.

type FeedbackValues = [FeedbackValue]

FeedbackValues hold information about one or more cells.

data Expression

Expression type gives us possible structures for cell input.

Constructors

Constant Integer

Cell Int Int

Add Expression Expression

Sub Expression Expression

Mult Expression Expression

Division Expression Expression

data ExError

ExError type can have a value of ParseError or EvaluationError or NoValue or ReferenceError.

Constructors

ParseError

EvaluationError

NoValue

ReferenceError