EXParser

Description

In module EXParser we export two main functions(evaluate and parseArithmetic) through which we come to the parsed expression, that is from input data-string to the output-some data structure like parse tree.

Documentation

table :: OperatorTable Char () Expression

table indicates possible operators.

factor :: Parser Expression

factor is parser for expression in combination of brackets.

number :: Parser Expression

number is parser for numbers.

cell :: Parser Expression

cell is for parsing notation for referencing other cells.

removeSpace :: String -> String

Function removeSpace is removing spaces.

expr :: Parser Expression

expr is parser for arithmetic expressions.

parseArithmetic

Arguments

:: String
as string we can take anything, meaning we can type in anything

-> Either ExError Expression
arithmetic expression don't return ExError

parseArithmetic is function for parsing, that is taking the actual string and returning type Either

evaluate :: Expression -> Either ExError Expression

This function evaluates expressions and then returns error or constant.

evaluateHelper :: Expression -> Either ExError Integer

evaluateHelper is function that helps us evaluate expression.