EXReactive

Documentation

makeGrid :: [[Element]] -> Element -> UI Element

We display a grid by using function makeGrid that takes a list of lists and Element and returns an UI Element.

getReference :: Expression -> Either ExError References

Function getReference gets cell references from Expression.

addToAccumulate :: FeedbackValues -> FeedbackValue -> FeedbackValues

Function addToAccumulate generates updated knowladge about other cells.

processExpression :: FeedbackValues -> Expression -> Either ExError Expression

This function generates reference cells in expression from current knowladge.

showOutput :: Either ExError Expression -> String

Function showOutput displays Information to the user that is in a cell.

toFeedback :: Coordinates -> Either ExError Expression -> FeedbackValue

Function toFeedback converts information to feedback information important for other cells.

bufferedEvent

Arguments

:: Element

-> UI (Event (Either ExError Expression), Handler (Either ExError Expression))
created new event

With function bufferedEvent we return on enter trigered event and event handler for error or expression.

makeAccumulator :: Event FeedbackValues -> UI (Behavior FeedbackValues)

Function makeAccumulator makes behavior for accumulating knowladge about other cells.

ioCell :: Event FeedbackValues -> Handler FeedbackValue -> Handler String -> Coordinates -> UI Element

Function ioCell configures and returns output cell for excel.

exprToStr :: Either ExError Expression -> String

Function exprToStr convertes Expression to a string.

setKeyIncrement :: Element -> Handler (Either ExError Expression) -> Behavior (Either ExError Expression) -> UI ()

Function configures increment/decrement with arrow keys.

displayElement :: Event String -> UI Element

Function generates displayed element.